+++ crates/hyperswitch_interfaces/src/webhooks.rs
@@ -1,0 +1,203 @@
+use common_utils::{crypto, errors::CustomResult, ext_traits::ValueExt};
+use error_stack::ResultExt;
+use masking::ExposeInterface;
+
+use crate::{api::ConnectorCommon, errors};
+pub struct IncomingWebhookRequestDetails<'a> {
+    pub method: http::Method,
+    pub uri: http::Uri,
+    pub headers: &'a http::header::HeaderMap,
+    pub body: &'a [u8],
+    pub query_params: String,
+}
+
+#[async_trait::async_trait]
+pub trait IncomingWebhook: ConnectorCommon + Sync {
+    fn get_webhook_body_decoding_algorithm(
+        &self,
+        _request: &IncomingWebhookRequestDetails<'_>,
+    ) -> CustomResult<Box<dyn crypto::DecodeMessage + Send>, errors::ConnectorError> {
+        Ok(Box::new(crypto::NoAlgorithm))
+    }
+
+    fn get_webhook_body_decoding_message(
+        &self,
+        request: &IncomingWebhookRequestDetails<'_>,
+    ) -> CustomResult<Vec<u8>, errors::ConnectorError> {
+        Ok(request.body.to_vec())
+    }
+
+    async fn decode_webhook_body(
+        &self,
+        request: &IncomingWebhookRequestDetails<'_>,
+        merchant_id: &str,
+        connector_webhook_details: Option<common_utils::pii::SecretSerdeValue>,
+        connector_name: &str,
+    ) -> CustomResult<Vec<u8>, errors::ConnectorError> {
+        let algorithm = self.get_webhook_body_decoding_algorithm(request)?;
+
+        let message = self
+            .get_webhook_body_decoding_message(request)
+            .change_context(errors::ConnectorError::WebhookBodyDecodingFailed)?;
+        let secret = self
+            .get_webhook_source_verification_merchant_secret(
+                merchant_id,
+                connector_name,
+                connector_webhook_details,
+            )
+            .await
+            .change_context(errors::ConnectorError::WebhookSourceVerificationFailed)?;
+
+        algorithm
+            .decode_message(&secret.secret, message.into())
+            .change_context(errors::ConnectorError::WebhookBodyDecodingFailed)
+    }
+
+    fn get_webhook_source_verification_algorithm(
+        &self,
+        _request: &IncomingWebhookRequestDetails<'_>,
+    ) -> CustomResult<Box<dyn crypto::VerifySignature + Send>, errors::ConnectorError> {
+        Ok(Box::new(crypto::NoAlgorithm))
+    }
+
+    async fn get_webhook_source_verification_merchant_secret(
+        &self,
+        merchant_id: &str,
+        connector_name: &str,
+        connector_webhook_details: Option<common_utils::pii::SecretSerdeValue>,
+    ) -> CustomResult<api_models::webhooks::ConnectorWebhookSecrets, errors::ConnectorError> {
+        let debug_suffix = format!(
+            "For merchant_id: {}, and connector_name: {}",
+            merchant_id, connector_name
+        );
+        let default_secret = "default_secret".to_string();
+        let merchant_secret = match connector_webhook_details {
+            Some(merchant_connector_webhook_details) => {
+                let connector_webhook_details = merchant_connector_webhook_details
+                    .parse_value::<api_models::admin::MerchantConnectorWebhookDetails>(
+                        "MerchantConnectorWebhookDetails",
+                    )
+                    .change_context_lazy(|| errors::ConnectorError::WebhookSourceVerificationFailed)
+                    .attach_printable_lazy(|| {
+                        format!(
+                            "Deserializing MerchantConnectorWebhookDetails failed {}",
+                            debug_suffix
+                        )
+                    })?;
+                api_models::webhooks::ConnectorWebhookSecrets {
+                    secret: connector_webhook_details
+                        .merchant_secret
+                        .expose()
+                        .into_bytes(),
+                    additional_secret: connector_webhook_details.additional_secret,
+                }
+            }
+
+            None => api_models::webhooks::ConnectorWebhookSecrets {
+                secret: default_secret.into_bytes(),
+                additional_secret: None,
+            },
+        };
+
+        //need to fetch merchant secret from config table with caching in future for enhanced performance
+
+        //If merchant has not set the secret for webhook source verification, "default_secret" is returned.
+        //So it will fail during verification step and goes to psync flow.
+        Ok(merchant_secret)
+    }
+
+    fn get_webhook_source_verification_signature(
+        &self,
+        _request: &IncomingWebhookRequestDetails<'_>,
+        _connector_webhook_secrets: &api_models::webhooks::ConnectorWebhookSecrets,
+    ) -> CustomResult<Vec<u8>, errors::ConnectorError> {
+        Ok(Vec::new())
+    }
+
+    fn get_webhook_source_verification_message(
+        &self,
+        _request: &IncomingWebhookRequestDetails<'_>,
+        _merchant_id: &str,
+        _connector_webhook_secrets: &api_models::webhooks::ConnectorWebhookSecrets,
+    ) -> CustomResult<Vec<u8>, errors::ConnectorError> {
+        Ok(Vec::new())
+    }
+
+    async fn verify_webhook_source(
+        &self,
+        request: &IncomingWebhookRequestDetails<'_>,
+        merchant_id: &str,
+        connector_webhook_details: Option<common_utils::pii::SecretSerdeValue>,
+        connector_name: &str,
+    ) -> CustomResult<bool, errors::ConnectorError> {
+        let algorithm = self
+            .get_webhook_source_verification_algorithm(request)
+            .change_context(errors::ConnectorError::WebhookSourceVerificationFailed)?;
+
+        let connector_webhook_secrets = self
+            .get_webhook_source_verification_merchant_secret(
+                merchant_id,
+                connector_name,
+                connector_webhook_details,
+            )
+            .await
+            .change_context(errors::ConnectorError::WebhookSourceVerificationFailed)?;
+
+        let signature = self
+            .get_webhook_source_verification_signature(request, &connector_webhook_secrets)
+            .change_context(errors::ConnectorError::WebhookSourceVerificationFailed)?;
+
+        let message = self
+            .get_webhook_source_verification_message(
+                request,
+                merchant_id,
+                &connector_webhook_secrets,
+            )
+            .change_context(errors::ConnectorError::WebhookSourceVerificationFailed)?;
+
+        algorithm
+            .verify_signature(&connector_webhook_secrets.secret, &signature, &message)
+            .change_context(errors::ConnectorError::WebhookSourceVerificationFailed)
+    }
+
+    fn get_webhook_object_reference_id(
+        &self,
+        _request: &IncomingWebhookRequestDetails<'_>,
+    ) -> CustomResult<api_models::webhooks::ObjectReferenceId, errors::ConnectorError>;
+
+    fn get_webhook_event_type(
+        &self,
+        _request: &IncomingWebhookRequestDetails<'_>,
+    ) -> CustomResult<api_models::webhooks::IncomingWebhookEvent, errors::ConnectorError>;
+
+    fn get_webhook_resource_object(
+        &self,
+        _request: &IncomingWebhookRequestDetails<'_>,
+    ) -> CustomResult<Box<dyn masking::ErasedMaskSerialize>, errors::ConnectorError>;
+
+    fn get_webhook_api_response(
+        &self,
+        _request: &IncomingWebhookRequestDetails<'_>,
+    ) -> CustomResult<services::api::ApplicationResponse<serde_json::Value>, errors::ConnectorError>
+    {
+        Ok(services::api::ApplicationResponse::StatusOk)
+    }
+
+    fn get_dispute_details(
+        &self,
+        _request: &IncomingWebhookRequestDetails<'_>,
+    ) -> CustomResult<crate::disputes::DisputePayload, errors::ConnectorError> {
+        Err(errors::ConnectorError::NotImplemented("get_dispute_details method".to_string()).into())
+    }
+
+    fn get_external_authentication_details(
+        &self,
+        _request: &IncomingWebhookRequestDetails<'_>,
+    ) -> CustomResult<crate::authentication::ExternalAuthenticationPayload, errors::ConnectorError>
+    {
+        Err(errors::ConnectorError::NotImplemented(
+            "get_external_authentication_details method".to_string(),
+        )
+        .into())
+    }
+}
