+++ crates/router/src/connector/cashtocode.rs
@@ -391,7 +391,7 @@ impl api::IncomingWebhook for Cashtocode {
             .get_webhook_source_verification_merchant_secret(
                 merchant_account,
                 connector_label,
-                merchant_connector_account,
+                merchant_connector_account.connector_webhook_details,
             )
             .await
             .change_context(errors::ConnectorError::WebhookSourceVerificationFailed)?;
