+++ crates/router/src/connector/signifyd.rs
@@ -697,7 +697,7 @@ impl api::IncomingWebhook for Signifyd {
             .get_webhook_source_verification_merchant_secret(
                 merchant_account,
                 connector_label,
-                merchant_connector_account,
+                merchant_connector_account.connector_webhook_details,
             )
             .await
             .change_context(errors::ConnectorError::WebhookSourceVerificationFailed)?;
