+++ crates/router/src/connector/zen.rs
@@ -617,7 +617,7 @@ impl api::IncomingWebhook for Zen {
             .get_webhook_source_verification_merchant_secret(
                 merchant_account,
                 connector_label,
-                merchant_connector_account,
+                merchant_connector_account.connector_webhook_details,
             )
             .await?;
         let signature =
