+++ crates/router/src/services/connector_integration_interface.rs
@@ -1,6 +1,8 @@
 use common_utils::{crypto, errors::CustomResult, request::Request};
 use hyperswitch_domain_models::{router_data::RouterData, router_data_v2::RouterDataV2};
-use hyperswitch_interfaces::connector_integration_v2::ConnectorIntegrationV2;
+use hyperswitch_interfaces::{
+    authentication::ExternalAuthenticationPayload, connector_integration_v2::ConnectorIntegrationV2,
+};
 
 use super::{BoxedConnectorIntegrationV2, ConnectorValidation};
 use crate::{
@@ -16,8 +18,8 @@ use crate::{
     types::{
         self,
         api::{
-            self, disputes, Connector, ConnectorV2, CurrencyUnit, ExternalAuthenticationPayload,
-            IncomingWebhookEvent, IncomingWebhookRequestDetails, ObjectReferenceId,
+            self, disputes, Connector, ConnectorV2, CurrencyUnit, IncomingWebhookEvent,
+            IncomingWebhookRequestDetails, ObjectReferenceId,
         },
         domain,
     },
@@ -99,25 +101,6 @@ impl api::IncomingWebhook for ConnectorEnum {
         }
     }
 
-    async fn get_webhook_body_decoding_merchant_secret(
-        &self,
-        db: &dyn StorageInterface,
-        merchant_id: &str,
-    ) -> CustomResult<Vec<u8>, errors::ConnectorError> {
-        match self {
-            Self::Old(connector) => {
-                connector
-                    .get_webhook_body_decoding_merchant_secret(db, merchant_id)
-                    .await
-            }
-            Self::New(connector) => {
-                connector
-                    .get_webhook_body_decoding_merchant_secret(db, merchant_id)
-                    .await
-            }
-        }
-    }
-
     fn get_webhook_body_decoding_message(
         &self,
         request: &IncomingWebhookRequestDetails<'_>,
@@ -133,16 +116,28 @@ impl api::IncomingWebhook for ConnectorEnum {
         db: &dyn StorageInterface,
         request: &IncomingWebhookRequestDetails<'_>,
         merchant_id: &str,
+        merchant_connector_account: domain::MerchantConnectorAccount,
+        connector_name: &str,
     ) -> CustomResult<Vec<u8>, errors::ConnectorError> {
         match self {
             Self::Old(connector) => {
                 connector
-                    .decode_webhook_body(db, request, merchant_id)
+                    .decode_webhook_body(
+                        request,
+                        merchant_id,
+                        merchant_connector_account,
+                        connector_name,
+                    )
                     .await
             }
             Self::New(connector) => {
                 connector
-                    .decode_webhook_body(db, request, merchant_id)
+                    .decode_webhook_body(
+                        request,
+                        merchant_id,
+                        merchant_connector_account,
+                        connector_name,
+                    )
                     .await
             }
         }
@@ -162,7 +157,7 @@ impl api::IncomingWebhook for ConnectorEnum {
         &self,
         merchant_account: &domain::MerchantAccount,
         connector_name: &str,
-        merchant_connector_account: domain::MerchantConnectorAccount,
+        connector_webhook_details: Option<common_utils::pii::SecretSerdeValue>,
     ) -> CustomResult<api_models::webhooks::ConnectorWebhookSecrets, errors::ConnectorError> {
         match self {
             Self::Old(connector) => {
@@ -170,7 +165,7 @@ impl api::IncomingWebhook for ConnectorEnum {
                     .get_webhook_source_verification_merchant_secret(
                         merchant_account,
                         connector_name,
-                        merchant_connector_account,
+                        connector_webhook_details,
                     )
                     .await
             }
@@ -179,7 +174,7 @@ impl api::IncomingWebhook for ConnectorEnum {
                     .get_webhook_source_verification_merchant_secret(
                         merchant_account,
                         connector_name,
-                        merchant_connector_account,
+                        connector_webhook_details,
                     )
                     .await
             }
@@ -216,40 +211,6 @@ impl api::IncomingWebhook for ConnectorEnum {
                 merchant_id,
                 connector_webhook_secrets,
             ),
-        }
-    }
-
-    async fn verify_webhook_source_verification_call(
-        &self,
-        state: &crate::routes::SessionState,
-        merchant_account: &domain::MerchantAccount,
-        merchant_connector_account: domain::MerchantConnectorAccount,
-        connector_name: &str,
-        request_details: &IncomingWebhookRequestDetails<'_>,
-    ) -> CustomResult<bool, errors::ConnectorError> {
-        match self {
-            Self::Old(connector) => {
-                connector
-                    .verify_webhook_source_verification_call(
-                        state,
-                        merchant_account,
-                        merchant_connector_account,
-                        connector_name,
-                        request_details,
-                    )
-                    .await
-            }
-            Self::New(connector) => {
-                connector
-                    .verify_webhook_source_verification_call(
-                        state,
-                        merchant_account,
-                        merchant_connector_account,
-                        connector_name,
-                        request_details,
-                    )
-                    .await
-            }
         }
     }
