+++ crates/router/src/types/api/authentication.rs
@@ -74,13 +74,6 @@ pub struct PostAuthenticationResponse {
     pub eci: Option<String>,
 }
 
-#[derive(Clone, serde::Deserialize, Debug, serde::Serialize, PartialEq, Eq)]
-pub struct ExternalAuthenticationPayload {
-    pub trans_status: common_enums::TransactionStatus,
-    pub authentication_value: Option<String>,
-    pub eci: Option<String>,
-}
-
 pub trait ConnectorAuthentication:
     services::ConnectorIntegration<
     Authentication,
