+++ crates/router/src/types/api/disputes.rs
@@ -1,3 +1,4 @@
+pub use hyperswitch_interfaces::disputes::DisputePayload;
 use masking::{Deserialize, Serialize};
 use time::PrimitiveDateTime;
 
@@ -11,20 +12,6 @@ pub struct DisputeId {
 pub use hyperswitch_domain_models::router_flow_types::dispute::{Accept, Defend, Evidence};
 
 pub use super::disputes_v2::{AcceptDisputeV2, DefendDisputeV2, DisputeV2, SubmitEvidenceV2};
-
-#[derive(Default, Debug)]
-pub struct DisputePayload {
-    pub amount: String,
-    pub currency: String,
-    pub dispute_stage: api_models::enums::DisputeStage,
-    pub connector_status: String,
-    pub connector_dispute_id: String,
-    pub connector_reason: Option<String>,
-    pub connector_reason_code: Option<String>,
-    pub challenge_required_by: Option<PrimitiveDateTime>,
-    pub created_at: Option<PrimitiveDateTime>,
-    pub updated_at: Option<PrimitiveDateTime>,
-}
 
 #[derive(Default, Debug, Deserialize, Serialize)]
 pub struct DisputeEvidence {
